backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
pydantic[email]==2.11.7
python-multipart==0.0.20
openpyxl==3.1.5
groq==0.31.0
python-dotenv==1.1.1
openai==1.102.0
